requirements.txt
# OCR with Kraken - Requirements
# Python 3.10 or 3.11 recommended (3.12 may have compatibility issues)
#
# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
#
# Step 1: Create virtual environment
#   python -m venv .venv
#   .venv\Scripts\activate       (Windows)
#   source .venv/bin/activate    (Linux/Mac)
#
# Step 2: Install PyTorch with CUDA (for GPU training)
#   Visit https://pytorch.org/get-started/locally/ and select your CUDA version
#   Example for CUDA 11.8:
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   Example for CUDA 12.1:
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   For CPU only:
#     pip install torch torchvision
#
# Step 3: Install remaining dependencies
#   pip install -r requirements.txt
#
# ============================================================================

# Core OCR library
kraken>=5.0.0,<7.0.0

# Image processing
Pillow>=9.0.0
scikit-image>=0.19.0

# Scientific computing
numpy>=1.21.0,<3.0.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Deep learning framework (for Kraken training)
# Note: PyTorch must be installed separately with CUDA support - see instructions above
pytorch-lightning>=2.0.0
lightning>=2.0.0

# Model format support (for .mlmodel files)
coremltools>=6.0

# Additional dependencies that may be needed
lxml>=4.6.0
regex>=2021.0.0
